==> RQ5/2-Statistics of reused posts(Tags).ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt


bug_df = pd.read_csv('./reuse post/so details.txt')
bug_df.info()
bug_df.head(2)

non_bug_df = pd.read_csv('./all post/so details.txt')
non_bug_df.info()
non_bug_df.head(2)



df_list = [bug_df, non_bug_df]
Tags = [[],[]]
for idx, temp_df in enumerate(df_list):
    for i in range(len(temp_df)):
        Tags[idx].extend(temp_df['Tags'][i].split(' '))



tag_dict_list = [dict(),dict()]
for idx, tag in enumerate(Tags):
    for t in tag:
        if t in tag_dict_list[idx].keys():
            tag_dict_list[idx][t] += 1
        else:
            tag_dict_list[idx][t] = 0


tag_dict1, tag_dict2 = tag_dict_list


sorted(tag_dict1.items(), key=lambda x: x[1],reverse=True)[1:10]


sorted(tag_dict2.items(), key=lambda x: x[1],reverse=True)[1:10]
