==> RQ5/4-High similarity reuse post analysis.ipynb <==
from datetime import datetime
import time
import os
from tqdm import tqdm
import pandas as pd
import re
import matplotlib.pyplot as plt
import numpy as np
import nltk
from nltk.corpus import stopwords
import pyLDAvis.gensim_models
import jieba.posseg as jp,jieba
from gensim.models.coherencemodel import CoherenceModel
from gensim.models.ldamodel import LdaModel
from gensim.corpora.dictionary import Dictionary
import warnings
warnings.filterwarnings("ignore")
import matplotlib.pyplot as plt
import matplotlib

# all_similarity = pd.read_csv('./all_similarity.txt')
# all_similarity.info()
#

# high_index = list(all_similarity[all_similarity['value']>=0.7].index)
#

# len(high_index)

all_sse = pd.read_csv('./all_sse.txt')
all_sse.columns=['value']
all_sse.info()


# all_similarity['value'].dropna()
all_sse['value'].dropna()



#%

np.mean(all_sse['value'].dropna())

all_sse = all_sse['value'].dropna().to_list()

len(all_sse)

pd.Series(all_sse).describe()


min_sse = min(all_sse)
max_sse = max(all_sse)
print(min_sse, max_sse)
def min_max(x):
    return (x-min_sse)/(max_sse-min_sse)

all_sse[:10]

all_sse = [min_max(i) for i in all_sse]
all_sse[:10]

pd.Series(all_sse).describe()

# count = [0]*5

# for idx, i in enumerate(all_similarity):
#     if 0<=i<0.2:
#         count[0] += 1
#     elif 0.2<=i<0.4:
#         count[1] += 1
#     elif 0.4<=i<0.6:
#         count[2] += 1
#     elif 0.6<=i<0.8:
#         count[3] += 1
#     elif 0.8<=i:
#         count[4] += 1
#
#
# count,sum(count)



# count = [0]*6
# for idx, i in enumerate(all_sse):
#     if 0<=i<1:
#         count[0] += 1
#     elif 1<=i<2:
#         count[1] += 1
#     elif 2<=i<3:
#         count[2] += 1
#     elif 3<=i<4:
#         count[3] += 1
#     elif 4<=i<5:
#         count[4] += 1
#     elif 5<=i<6:
#         count[5] += 1



# count = [0]*3
# for idx, i in enumerate(all_sse):
#     if 0<=i<2:
#         count[0] += 1
#     elif 2<=i<4:
#         count[1] += 1
#     elif 4<=i<6:
#         count[2] += 1


count = [0]*5
for idx, i in enumerate(all_sse):
    if 0<=i<0.2:
        count[0] += 1
    elif 0.2<=i<0.4:
        count[1] += 1
    elif 0.4<=i<0.6:
        count[2] += 1
    elif 0.6<=i<0.8:
        count[3] += 1
    elif 0.8<=i:
        count[4] += 1

print(count,sum(count))
print([(i/sum(count))*100 for i in count])

with open('./commit_name.txt','r', encoding='utf-8') as f:
    lines = f.readlines()
commit_name = [i.rstrip('\n') for i in lines]

commit_name.index('D-CCFinder-NewData-javamelody-10-14717.java.txt')

commit_name.index('D-CCFinder-NewData-algs4-4-5288.java.txt')

len(all_sse), all_sse.iloc[58]

def min_max(x):
    return (x-min_sse)/(max_sse-min_sse)
all_sse = pd.read_csv('./all_sse.txt')
all_sse.columns=['value']
all_sse.info()
all_sse['value'] = all_sse['value'].apply(lambda x: min_max(x))
high_index = list(all_sse[(all_sse['value']>=0.6) & (all_sse['value']<0.8)].index)


len(high_index)

all_sse.describe()

import shutil
src_path = 'D:/CCFinder/Commit2So(A commit will correspond to multiple sos)/'
trg_path = './Multiplexed post groups with sse greater than or equal to 0.6 and less than 0.8/'

for i in tqdm(high_index):
    file = commit_name[i]
    shutil.copyfile(src_path+file, trg_path+file)


files = os.listdir(trg_path)
for file in tqdm(files):
    with open(trg_path+file, 'r', encoding='utf-8') as f:
        lines = f.readlines()
        lines = set(lines)
    with open(trg_path+file, 'w', encoding='utf-8') as f:
        for line in lines:
            f.write(line)


for file in tqdm(files):
    with open(trg_path+file, 'r', encoding='utf-8') as f:
        lines = f.readlines()
        lines = set(lines)
        if len(lines)>=2 and len(lines)<3:
            print(file,len(lines))
